New Development: No             JIRA Issue: Yes (#522)         Ready for Test: Yes      Interface Changes: Yes What Interface Changed: The scaling factor is changed from double to float          Test Programs:   Put in Release Notes: No              Module(s): sdms            Description: changed the scaling factor in sdms.scale to float.
git-svn-id: https://svn.cv.nrao.edu/svn/casa/branches/project-singledish-ms@30970 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ gcwrap/tools/singledish/sdms_cmpt.cc
@@ -94,7 +94,7 @@ sdms::name()
 }
 
 bool
-sdms::scale(double const factor)
+sdms::scale(float const factor)
 {
   bool rstat(false);
   try {
